Add explicit integer casts
WebKit uses the -Wshorten-64-to-32 flag which warns on these cases.

Bug: 3439

+++ src/libANGLE/renderer/vulkan/RenderTargetVk.cpp
@@ -38,8 +38,8 @@ RenderTargetVk::RenderTargetVk(RenderTargetVk &&other)
 void RenderTargetVk::init(vk::ImageHelper *image,
                           vk::ImageView *imageView,
                           vk::ImageView *cubeImageFetchView,
-                          size_t levelIndex,
-                          size_t layerIndex)
+                          uint32_t levelIndex,
+                          uint32_t layerIndex)
 {
     mImage              = image;
     mImageView          = imageView;
@@ -131,7 +131,7 @@ const vk::Format &RenderTargetVk::getImageFormat() const
 gl::Extents RenderTargetVk::getExtents() const
 {
     ASSERT(mImage && mImage->valid());
-    return mImage->getLevelExtents2D(mLevelIndex);
+    return mImage->getLevelExtents2D(static_cast<uint32_t>(mLevelIndex));
 }
 
 void RenderTargetVk::updateSwapchainImage(vk::ImageHelper *image, vk::ImageView *imageView)
@@ -185,8 +185,9 @@ angle::Result RenderTargetVk::flushStagedUpdates(ContextVk *contextVk)
 
     vk::CommandBuffer *commandBuffer;
     ANGLE_TRY(mImage->recordCommands(contextVk, &commandBuffer));
-    return mImage->flushStagedUpdates(contextVk, mLevelIndex, mLevelIndex + 1, mLayerIndex,
-                                      mLayerIndex + 1, commandBuffer);
+    return mImage->flushStagedUpdates(
+        contextVk, static_cast<uint32_t>(mLevelIndex), static_cast<uint32_t>(mLevelIndex + 1),
+        static_cast<uint32_t>(mLayerIndex), static_cast<uint32_t>(mLayerIndex + 1), commandBuffer);
 }
 
 }  // namespace rx
